refactor(058): reorganize puzzle domain
