+++ app/src/main/java/org/cescfe/numpairs/domain/puzzle/Puzzle.kt
@@ -7,10 +7,8 @@ data class Puzzle(val board: Board, val strip: Strip) {
     val completionState: PuzzleCompletionState
         get() = when {
             isIncomplete -> PuzzleCompletionState.INCOMPLETE
-            board.tiles.any { tile -> tile.resolutionState == TileResolutionState.INCORRECT } ->
-                PuzzleCompletionState.INCORRECT_TILES
-            resolvedTileAssignments().size != board.tiles.size ->
-                PuzzleCompletionState.MISSING_STRIP_ENTRY_IDENTITIES
+            hasIncorrectTiles -> PuzzleCompletionState.INCORRECT_TILES
+            hasMissingResolvedTileAssignments -> PuzzleCompletionState.MISSING_STRIP_ENTRY_IDENTITIES
             hasMismatchedSumProductPairings -> PuzzleCompletionState.MISMATCHED_SUM_PRODUCT_PAIRINGS
             hasInvalidStripEntryUsage -> PuzzleCompletionState.INVALID_STRIP_ENTRY_USAGE
             else -> PuzzleCompletionState.SOLVED
